+++ algorithms/algorithms.h
@@ -10,28 +10,28 @@
 
 using namespace std;
 
-pair<long long, int> selectionSort(vector<int>&);
+pair<long long, long long> selectionSort(vector<int>&);
 
-pair<long long, int> insertionSort(vector<int>&);
+pair<long long, long long> insertionSort(vector<int>&);
 
-pair<long long, int> binaryInsertionSort(vector<int>&);
+pair<long long, long long> binaryInsertionSort(vector<int>&);
 
-pair<long long, int> bubbleSort(vector<int>&);
+pair<long long, long long> bubbleSort(vector<int>&);
 
-pair<long long, int> shakerSort(vector<int>&);
+pair<long long, long long> shakerSort(vector<int>&);
 
-pair<long long, int> shellSort(vector<int>&);
+pair<long long, long long> shellSort(vector<int>&);
 
-pair<long long, int> heapSort(vector<int>&);
+pair<long long, long long> heapSort(vector<int>&);
 
-pair<long long, int> mergeSort(vector<int>&);
+pair<long long, long long> mergeSort(vector<int>&);
 
-pair<long long, int> quickSort(vector<int>&);
+pair<long long, long long> quickSort(vector<int>&);
 
-pair<long long, int> countingSort(vector<int>&);
+pair<long long, long long> countingSort(vector<int>&);
 
-pair<long long, int> radixSort(vector<int>&);
+pair<long long, long long> radixSort(vector<int>&);
 
-pair<long long, int> flashSort(vector<int>&);
+pair<long long, long long> flashSort(vector<int>&);
 
-pair<long long, int> sortDataByAlgorithm(string, vector<int>&);
+pair<long long, long long> sortDataByAlgorithm(string, vector<int>&);
